iOS SDK China Translation for Request Button
Summary:
Added localization files and reorganized resources into a single bundle as follows:

UberRidesResources.bundle
- Media.xcassets
- *.lproj/

where there is a *.lproj for every supported language. The podspec resources path was updated to use this bundle. In this revision, zh-Hans (simplified Chinese) and zh-Hant (traditional Chinese) were added. Currently, both are using the same translated string of "Ride there with Uber". Once we have a translated string for traditional Chinese, this can be changed in the appropriate localization file.

There is a .strings file for every supported language (in the respective *.lproj folder) which contains a list of key-value pairs. The key is what is used in the SDK, and the value is different according to the language. One key was added in this revision for the request button title text.

{F725101}

{F725102}

Test Plan:
- Create a test app that uses the UberRides SDK and add the request button to a view
-- Run app with default system language (English) and verify button says "Ride there with Uber"
-- Run app with system language as Chinese (simplified) and verify button says "搭乘优步前往"
-- Run app with system language as Chinese (traditional) and verify button says "搭乘优步前往"
-- Run app with system language as Chinese (traditional, Hong Kong) and verify button says "搭乘优步前往"
- Verify button visually in both languages

Reviewers: #external_api, chkim

Reviewed By: #external_api, chkim

Subscribers: chrissaad

Differential Revision: https://code.uberinternal.com/D289992

Projects: #external_api