[iOS SDK] Fix for Github Issue #6 (pickupLocation is ignored)
Summary:
[[ https://github.com/uber/rides-ios-sdk/issues/6 | Link to Github Issue ]]

This issue was caused due to the deeplink being built too early. There were 5 calls to build():
- setDeepLink(); called upon the initialization of the RequestButton. At this point, since there are no coordinates set, there would be a 'pickup=my_location' added in the parameters
- setProductId(); if this is set before the coordinates, the above behaviour would occur
- setPickupLocationToCurrentLocation(), setPickupLocation(), and setDropoffLocation()

In this revision, the issue was fixed through removing all the above calls, and adding one call to build() in the tap event handler of the RequestButton. This means that instead of incrementally building the deeplink, it is now built once when the button is tapped. Initial timing tests indicate that this solution has a negligible effect on the execution time when the request button is tapped and the deeplink is built.

This revision also switches to using NSURLComponents to build the deeplink instead of string formatting.

The following screenshots are a result of the fixes made in this revision, for both the sample code provided in the Github issue and our own example app in the SDK repository.

{F727637}

{F727638}

Test Plan:
- Executed code provided by issue reporter and verified that pickup location is the one provided (in Uruguay) and not current location
- Executed Swift SDK Example app and verified that pickup and dropoff location are those provided and not current location
- Set no pickup or dropoff location and verified that the pickup location is set to current location

Reviewers: jbrophy, #external_api, chkim

Reviewed By: #external_api, chkim

Differential Revision: https://code.uberinternal.com/D291174

Projects: #external_api