[iOS SDK] Use dictionary instead of array for ride-parameters
Summary:
Previously, ride parameters were added to an array; thus, old keys wouldn't be replaced with new values. This revision changes the structure to use a private QueryParameters class that internally handles the mapping of the parameters. It has functions to setParameter, deleteParameters, and check if a parameter exists in the dictionary.

Additionally, when setPickupLocationToCurrentLocation() is called, the parameters for latitude, longitude, address, and nickname are removed. When setPickupLocation() is called, the parameter for default location is removed. This is so that if a user calls both functions, both sets aren't included in the deep link.

Lastly, a small optimization was made to avoid rebuilding the deep link on multiple taps of the button. The QueryParameters class has a pendingChanges flag that is set to true whenever a parameter is set or removed, and set to false once the deep link has been built.

Test Plan:
Basic functionality
- No parameters set: verify deep link sets location to current
- Set pickup location: verify deep link sets pickup location appropriately
- Set pickup and drop-off locations: verify deep link sets pickup and drop-off locations appropriately

Overriding parameters
- Call setPickupLocation twice with different locations: verify deep link only has parameters for the latter location
- Call setPickupLocation, then setPickupLocationToCurrentLocation: verify deep link only has parameters for current location
- Call setPickupLocationToCurrentLocation, then setPickupLocation: verify deep link only has parameters for pickup location that was set (ie. that there is no pickup=my_location)

Deep Link Building
- Set breakpoint in build() function
-- Tap button once and verify deep link is built
-- Make no changes, tap button for a second time: verify that deep link is not built again
-- Make change (setParameter): verify deep link is built again
-- Make change (deleteParameter): verify deep link is built again

Reviewers: chkim, #external_api, jbrophy

Reviewed By: jbrophy

Subscribers: texter

Differential Revision: https://code.uberinternal.com/D291660

Projects: #external_api