Fixed a bug where the system dialog box:
"<App Name>" wants to open "Uber"
Cancel / Open

would cause SSO to fail because the system dialog triggers the UIApplicationDidBecomeActive notification.
The fix is to make sure that the notification is preceded by a UIApplicationWillEnterForeground notification,
which happens only when the there's been a context switch to the Uber app.

Fixed a related bug where, if the user were to select Cancel in the system dialog box (above), the library
would attempt to use fallback to a web view, when the correct handling would be to stop attempting to log
in and return an error to the caller.